Refactored profile details to pull out functionality for grabbing user details into a utility function and generally refactoring code
